feat: split App Store into standalone Discover and Library pages
Discover (/apps) and Library (/apps/library) become independent
path-addressable pages with their own sidebar entries, replacing the
in-page tab toggle. Shared registry/install-state logic moves into a
single useAppsData hook so the two pages cannot drift. The static
library segment registers ahead of the /apps/:name catch-all and
'library' joins the reserved app names (code: reserved_app_name).
Legacy sessionStorage appstore-tab values migrate via a one-time
redirect.
