+++ website/playwright/apps.spec.ts
@@ -1,54 +1,18 @@
-import { test, expect, type APIRequestContext, type Page } from '@playwright/test'
-
 /**
- * Apps routes e2e — covers /apps, /apps/detail/:name, /apps/:name.
+ * E2E: the split App Store surfaces.
  *
- * The Apps page is a hybrid storefront (upstream #532): a "Discover" tab with an
- * editorial layer (FeaturedSpotlight blocks) over a category-railed
- * catalog, and a "Library" tab managing installed apps. It replaced the earlier
- * Installed/Browse tabs.
+ * Discover (/apps) and Library (/apps/library) are standalone pages (the
+ * hybrid tabbed AppsPage and its SegmentedControl are gone): Discover is the
+ * storefront (editorial layer + category rail + catalog), Library manages
+ * installed apps. The static `library` segment registers ahead of the
+ * /apps/:name installed-app catch-all and is a reserved app name server-side.
  *
- * Harness state:
- * - 1 installed & enabled builtin: "Task Runner" (name: "projects", route /projects)
+ * Preconditions provided by the stub gateway:
  * - Several disabled builtins populate the Discover catalog
- * - Discover is the default tab on a fresh session (initialTab(), AppsPage.tsx:46)
- *
- * Tests assert the real harness state unconditionally. No conditional branches.
- *
- * Direct navigation to /apps/detail/:name works: the server's SPA fallback
- * excludes only the two sub-namespaces apps/routes.py serves
- * (/apps/{name}/api/ and /apps/{name}/ui/), so every other /apps/ path reaches
- * React Router. It used to exclude any /apps/{seg}/ path, reading "detail" as
- * the app name and 404ing the bare URL, which is why these tests previously
- * navigated via pushState instead. gotoDetail() now hits the URL directly, so a
- * regression re-appears here as a 404 rather than being routed around.
+ * - Task Runner (name: "projects") is installed and enabled
  */
 
-type AppEntry = {
-  name: string
-  displayName: string
-  enabled: boolean
-  origin?: string
-  manifest?: { description?: string; ui?: { pages?: { route: string }[] } }
-}
-
-/** Fetch the installed apps list from the API. */
-async function listApps(request: APIRequestContext): Promise<AppEntry[]> {
-  const res = await request.get('/api/apps')
-  expect(res.ok()).toBeTruthy()
-  return await res.json()
-}
-
-/**
- * Locate a tab segment. SegmentedControl exposes no aria-label, role, or
- * data-testid -- only title={label} -- so title is the sole stable handle. It is
- * also the only one that survives both hazards: compact mode drops the label
- * text from the DOM for inactive segments, and the Library segment renders a
- * count badge inside the button, which makes its accessible name "Library 1".
- */
-function tab(page: Page, label: 'Discover' | 'Library') {
-  return page.locator(`button[title="${label}"]`)
-}
+import { test, expect, type Page } from '@playwright/test'
 
 /**
  * Catalog cards. The same aria-label is emitted by FeaturedSpotlight AND
@@ -64,19 +28,26 @@ function browseCards(page: Page) {
  * testid or aria-label, but it renders the app name as a real <button> wired to
  * onDetail -- so role-scope it rather than matching loose text.
  *
- * Scoped to the #main-content landmark (App.tsx:1984): an installed builtin also
- * appears in the nav rail, so an unscoped role+name lookup resolves to 2
- * elements. This mirrors the convention capabilities.spec.ts already uses.
+ * Scoped to the #main-content landmark: an installed builtin also appears in
+ * the nav rail, so an unscoped role+name lookup resolves to 2 elements. This
+ * mirrors the convention capabilities.spec.ts already uses.
  */
 function libraryCard(page: Page, displayName: string) {
   return page.locator('#main-content').getByRole('button', { name: displayName, exact: true })
 }
 
-async function gotoApps(page: Page) {
+async function gotoDiscover(page: Page) {
   await page.goto('/apps', { waitUntil: 'domcontentloaded' })
-  // Ready-signal is the tab segment's title attribute, not the page subtitle:
-  // subtitle copy is prose a designer can reword, title is a stable handle.
-  await expect(tab(page, 'Discover')).toBeVisible({ timeout: 10000 })
+  // Ready-signal is the catalog heading (category === 'all'): stable structure,
+  // not subtitle prose a designer can reword.
+  await expect(page.getByRole('heading', { name: 'All apps' })).toBeVisible({ timeout: 10000 })
+}
+
+async function gotoLibrary(page: Page) {
+  await page.goto('/apps/library', { waitUntil: 'domcontentloaded' })
+  // The Library search box is the page's stable ready-signal (its aria-label
+  // is page-specific after the split).
+  await expect(page.getByRole('textbox', { name: 'Search library' })).toBeVisible({ timeout: 10000 })
 }
 
 /**
@@ -92,67 +63,60 @@ async function gotoDetail(page: Page, appName: string) {
   await expect(page.getByRole('button', { name: 'Back to Apps' })).toBeVisible({ timeout: 10000 })
 }
 
-test.describe('Apps Page — /apps', () => {
-  test('renders the header, both tab segments, and the Discover catalog', async ({ page }) => {
-    await gotoApps(page)
-    await expect(tab(page, 'Discover')).toBeVisible({ timeout: 5000 })
-    await expect(tab(page, 'Library')).toBeVisible()
-    // Discover's two-column layout: the "All apps" heading (category === 'All'),
-    // the CategoryRail, and the catalog's sort control.
-    //
+test.describe('Discover Page — /apps', () => {
+  test('renders the storefront: catalog heading, category rail, sort control', async ({ page }) => {
+    await gotoDiscover(page)
     // Deliberately NOT asserting the "N apps" count line: that exact string is
     // rendered twice -- once by the CategoryRail as its source total, once as the
     // catalog result count -- so the locator is ambiguous, and .first() would
     // silently assert the rail's total instead of the catalog's.
-    await expect(page.getByRole('heading', { name: 'All apps' })).toBeVisible({ timeout: 5000 })
     await expect(page.getByRole('button', { name: 'Add source' })).toBeVisible()
     await expect(page.getByRole('combobox', { name: 'Sort apps' })).toBeVisible()
   })
 
-  test('Discover is the default tab on a fresh session', async ({ page }) => {
-    // initialTab() (AppsPage.tsx:46) returns 'discover' when sessionStorage has
-    // no persisted tab. Playwright's storageState does not carry sessionStorage,
-    // so every test starts fresh -- this is deterministic, not incidental.
-    await gotoApps(page)
-    await expect(page.getByRole('heading', { name: 'All apps' })).toBeVisible({ timeout: 5000 })
-    await expect(browseCards(page).first()).toBeVisible({ timeout: 10000 })
-  })
-
-  test('Discover renders catalog cards for the available builtins', async ({ page }) => {
-    await gotoApps(page)
+  test('renders catalog cards for the available builtins', async ({ page }) => {
+    await gotoDiscover(page)
     const cards = browseCards(page)
     await expect(cards.first()).toBeVisible({ timeout: 10000 })
     expect(await cards.count()).toBeGreaterThan(0)
   })
 
-  test('Discover search filters the catalog down to its empty state', async ({ page }) => {
-    await gotoApps(page)
+  test('search filters the catalog down to its empty state', async ({ page }) => {
+    await gotoDiscover(page)
     await expect(browseCards(page).first()).toBeVisible({ timeout: 10000 })
 
     const search = page.getByRole('textbox', { name: 'Search apps' })
     await search.fill('zzz_no_match_xyz')
 
     // A non-empty query also clears the editorial layer (showEditorial requires
-    // !query.trim(), AppsPage.tsx:207), so every card unmounts -- not just the
-    // AppListRows.
+    // !query.trim()), so every card unmounts -- not just the AppListRows.
     await expect(page.getByTestId('empty-state-title')).toHaveText('No matching apps', { timeout: 5000 })
     await expect(browseCards(page)).toHaveCount(0)
   })
 
-  test('Library tab lists the installed Task Runner app', async ({ page }) => {
-    await gotoApps(page)
-    await tab(page, 'Library').click()
+  test('a stored legacy library tab redirects /apps to /apps/library once', async ({ page }) => {
+    // The pre-split page persisted its tab in sessionStorage; DiscoverPage
+    // translates a stored library value into a one-shot replace-redirect and
+    // clears the key. addInitScript runs before the app boots, mirroring a
+    // user upgrading with the old value persisted.
+    await page.addInitScript(() => sessionStorage.setItem('appstore-tab', 'library'))
+    await page.goto('/apps', { waitUntil: 'domcontentloaded' })
+    await page.waitForURL('**/apps/library', { timeout: 10000 })
+    await expect(page.getByRole('textbox', { name: 'Search library' })).toBeVisible({ timeout: 10000 })
+  })
+})
+
+test.describe('Library Page — /apps/library', () => {
+  test('lists the installed Task Runner app', async ({ page }) => {
+    await gotoLibrary(page)
     await expect(libraryCard(page, 'Task Runner')).toBeVisible({ timeout: 10000 })
   })
 
-  test('Library search narrows to matching installed apps', async ({ page }) => {
-    await gotoApps(page)
-    await tab(page, 'Library').click()
+  test('search narrows to matching installed apps', async ({ page }) => {
+    await gotoLibrary(page)
     await expect(libraryCard(page, 'Task Runner')).toBeVisible({ timeout: 10000 })
 
-    // One SearchInput serves both tabs; its placeholder switches per tab but the
-    // aria-label stays "Search apps".
-    const search = page.getByRole('textbox', { name: 'Search apps' })
+    const search = page.getByRole('textbox', { name: 'Search library' })
     await search.fill('zzz_no_match_xyz')
     await expect(page.getByTestId('empty-state-title')).toHaveText('No matching apps', { timeout: 5000 })
 
@@ -160,15 +124,15 @@ test.describe('Apps Page — /apps', () => {
     await expect(libraryCard(page, 'Task Runner')).toBeVisible({ timeout: 5000 })
   })
 
-  test('tab round-trip returns to the Discover catalog', async ({ page }) => {
-    await gotoApps(page)
+  test('page round-trip: Library and Discover are independently routable', async ({ page }) => {
+    await gotoDiscover(page)
     await expect(page.getByRole('heading', { name: 'All apps' })).toBeVisible({ timeout: 5000 })
 
-    await tab(page, 'Library').click()
+    await gotoLibrary(page)
     await expect(libraryCard(page, 'Task Runner')).toBeVisible({ timeout: 10000 })
     await expect(page.getByRole('heading', { name: 'All apps' })).toHaveCount(0)
 
-    await tab(page, 'Discover').click()
+    await gotoDiscover(page)
     await expect(page.getByRole('heading', { name: 'All apps' })).toBeVisible({ timeout: 5000 })
   })
 })
@@ -188,8 +152,7 @@ test.describe('App Detail Page — /apps/detail/:name', () => {
   })
 
   test('navigating from a Discover catalog card reaches the detail page', async ({ page }) => {
-    await gotoApps(page)
-    // Discover is the default tab, so the catalog is already on screen.
+    await gotoDiscover(page)
     const firstCard = browseCards(page).first()
     await expect(firstCard).toBeVisible({ timeout: 10000 })
     await firstCard.click()
@@ -228,5 +191,12 @@ test.describe('App Page — /apps/:name', () => {
     await expect(page.locator('text=is not installed')).toBeVisible({ timeout: 10000 })
     const url = page.url()
     expect(url).not.toContain('/chat')
+  })
+
+  test('the static /apps/library segment wins over the /apps/:name catch-all', async ({ page }) => {
+    // Registration order is the contract this pins from the browser side; the
+    // server-side half is the reserved-name refusal (reserved_app_name).
+    await gotoLibrary(page)
+    await expect(page.locator('text=is not installed')).toHaveCount(0)
   })
 })
