+++ website/src/components/commandPalette/providers/pagesProvider.ts
@@ -12,6 +12,7 @@ import {
   Bot,
   Server,
   LayoutGrid,
+  Compass,
 } from 'lucide-react'
 
 import { getAdvertisedSurfaces, surfaceLabel } from '../../../surfaces/registry'
@@ -86,7 +87,10 @@ const EXTRA_PAGES: readonly (Omit<PageEntry, 'title'> & { previewFlag?: string }
   // The App Store surface is `hiddenFromNav` (it renders as the Apps-header
   // "Explore" accent link, not a rail row), so it must be listed here to
   // stay reachable from the palette.
-  { key: 'apps', route: '/apps', icon: inlineIcon(LayoutGrid) },
+  { key: 'apps', route: '/apps', icon: inlineIcon(Compass) },
+  // Library is its own page after the App Store split; the rail row exists,
+  // but the palette resolves entries from this list, so it needs its own row.
+  { key: 'apps-library', route: '/apps/library', icon: inlineIcon(LayoutGrid) },
   // Inbound webhooks is `hiddenFromNav` too (reached from Settings → Webhooks),
   // so the registry no longer offers it and the palette needs it from here. It
   // is ALSO preview-gated, so it carries `previewFlag` and stays out of the
@@ -115,7 +119,11 @@ const EXTRA_PAGES: readonly (Omit<PageEntry, 'title'> & { previewFlag?: string }
  * `dynamic-keys-baseline.json` — a ratchet that only goes down.
  */
 const EXTRA_PAGE_TITLE_KEY: Record<string, string> = {
-  apps: 'components.commandPalette.providers.pagesProvider.explore',
+  // Reuses the sidebar's own labels so the palette and the rail cannot
+  // disagree on what the pages are called (the pre-split "Explore" title
+  // survived the rail's rename to Discover exactly this way).
+  apps: 'nav.discover',
+  'apps-library': 'nav.library',
   // Reuses strings that already exist in every catalog rather than adding new
   // ones. Titled "Inbound webhooks", not "Webhooks", to stay distinguishable
   // from the `hooks` entry (the agent-hooks page) that sits beside it.
