+++ website/src/i18n/locales/en-XA.json
@@ -13,8 +13,6 @@
     "discord_community": "[Ðìşçøŕð çøɱɱùñìţý ···············]",
     "dismiss": "[Ðìşɱìşş ···········]",
     "dsk": "[ÐŞĶ ·····]",
-    "explore": "[Èẋþĺøŕè ···········]",
-    "explore_apps": "[Èẋþĺøŕè Àþþş ···········]",
     "k_search_for_anything": "[⌘Ķ — Şèàŕçĥ ƒøŕ àñýţĥìñğ… ··················]",
     "kiro_credit_usage_checking": "[Ķìŕø çŕèðìţ ùşàğè — çĥèçķìñğ… ····················]",
     "kiro_credit_usage_checking_2": "[Ķìŕø çŕèðìţ ùşàğè — çĥèçķìñğ ····················]",
@@ -5155,7 +5153,6 @@
         },
         "pagesProvider": {
           "developer": "[Ðèṽèĺøþèŕ ··············]",
-          "explore": "[Èẋþĺøŕè ···········]",
           "hooks": "[Ĥøøķş ········]",
           "kirocrew_agents": "[{{productName}} Àğèñţş ···········]",
           "logs": "[Ĺøğş ······]",
@@ -7524,12 +7521,11 @@
       "loading_app": "[Ĺøàðìñğ àþþ… ···········]"
     },
     "appsPage": {
-      "add_an_app_source_gear_icon_above_or_install_fro": "[Àðð àñ àþþ şøùŕçè (ğèàŕ ìçøñ àƀøṽè) øŕ ìñşţàĺĺ ƒŕøɱ à ĺøçàĺ þàţĥ. ·······················]",
+      "add_an_app_source_gear_icon_above_or_install_fro": "[Àðð àñ àþþ şøùŕçè (Şøùŕçèş, ţøþ ŕìğĥţ) øŕ ìñşţàĺĺ ƒŕøɱ à ĺøçàĺ þàţĥ. ························]",
       "agent_one": "[{{count}} àğèñţ ·········]",
       "agent_other": "[{{count}} àğèñţş ···········]",
       "app_one": "[{{count}} àþþ ······]",
       "app_other": "[{{count}} àþþş ········]",
-      "apps": "[Àþþş ······]",
       "available": "[àṽàìĺàƀĺè ··············]",
       "cancel": "[Çàñçèĺ ·········]",
       "category": "[Çàţèğøŕý ············]",
@@ -7537,7 +7533,6 @@
       "cron_job_one": "[{{count}} çŕøñ ĵøƀ ··············]",
       "cron_job_other": "[{{count}} çŕøñ ĵøƀş ···············]",
       "dependencies": "[Ðèþèñðèñçìèş: ············]",
-      "discover_install_and_manage_agentic_apps": "[Ðìşçøṽèŕ, ìñşţàĺĺ, àñð ɱàñàğè àğèñţìç àþþş ·····················]",
       "dismiss_message": "[Ðìşɱìşş ɱèşşàğè ··············]",
       "keep_app_data": "[Ķèèþ àþþ ðàţà ············]",
       "keep_this_dependency": "[Ķèèþ ţĥìş ðèþèñðèñçý ··················]",
@@ -7561,23 +7556,23 @@
       "update_one": "[{{count}} ùþðàţè ···········]",
       "update_other": "[{{count}} ùþðàţèş ············]",
       "v": "[ṽ ··]",
-      "installed_app_find_in_library_and_enable": "[Ìñşţàĺĺèð {{name}}. Ţùŕñ ìţ øñ ƀèĺøẁ ţø èñàƀĺè ìţ. ·····················]",
+      "installed_app_find_in_library_and_enable": "[Ìñşţàĺĺèð {{name}}. Èñàƀĺè ìţ ìñ Ĺìƀŕàŕý. ·················]",
       "synced_from_its_source_directory": "[Şýñçèð {{name}} ƒŕøɱ ìţş şøùŕçè ðìŕèçţøŕý. ·················]",
       "updated_from_the_registry": "[Ùþðàţèð {{name}} ƒŕøɱ ţĥè ŕèğìşţŕý. ···················]",
       "action_failed": "[Ƒàìĺèð ţø {{action}} {{name}} ··········]",
       "all_apps": "[Àĺĺ àþþş ············]",
       "built_in_kirocrew": "[Ɓùìĺţ-ìñ · ķìŕøçŕèẁ ·················]",
-      "discover": "[Ðìşçøṽèŕ ············]",
       "failed_to_enable": "[Ƒàìĺèð ţø èñàƀĺè {{name}} ···············]",
       "failed_to_load_apps": "[Ƒàìĺèð ţø ĺøàð àþþş ·················]",
       "failed_to_load_registry": "[Ƒàìĺèð ţø ĺøàð ŕèğìşţŕý ················]",
       "failed_to_uninstall": "[Ƒàìĺèð ţø ùñìñşţàĺĺ {{name}} ··················]",
       "failed_to_update": "[Ƒàìĺèð ţø ùþðàţè: {{names}} ················]",
-      "find_apps_in_the_discover_tab_or_install_from_a": "[Ƒìñð àþþş ìñ ţĥè Ðìşçøṽèŕ ţàƀ, øŕ ìñşţàĺĺ ƒŕøɱ à ĺøçàĺ þàţĥ ṽìà ţĥè şøùŕçèş ğèàŕ àƀøṽè. ··························]",
-      "hidden_you_can_re_enable_it_from_the_discover_ta": "[Ðìşàƀĺèð. Ýøù çàñ ŕè-èñàƀĺè ìţ ƒŕøɱ ţĥè Ðìşçøṽèŕ ţàƀ. ···················]",
+      "find_apps_in_the_discover_tab_or_install_from_a": "[Ƒìñð àþþş øñ ţĥè Ðìşçøṽèŕ þàğè, øŕ ìñşţàĺĺ ƒŕøɱ à ĺøçàĺ þàţĥ ṽìà Şøùŕçèş ţĥèŕè. ························]",
+      "hidden_you_can_re_enable_it_from_the_discover_ta": "[Ðìşàƀĺèð. Ýøù çàñ ŕè-èñàƀĺè ìţ ƒŕøɱ ţĥè Ðìşçøṽèŕ þàğè. ···················]",
       "keep_dependency": "[Ķèèþ ðèþèñðèñçý {{name}} ··············]",
       "kirocrew_registry": "[{{productName}} ŕèğìşţŕý ··············]",
       "library": "[Ĺìƀŕàŕý ···········]",
+      "manage_your_installed_apps": "[Ṁàñàğè ýøùŕ ìñşţàĺĺèð àþþş ··················]",
       "no_apps_installed_yet": "[Ñø àþþş ìñşţàĺĺèð ýèţ ···············]",
       "removing": "[Ŕèɱøṽìñğ… ··············]",
       "search_library": "[Şèàŕçĥ ĺìƀŕàŕý ·············]",
@@ -12540,6 +12535,10 @@
       "waiting_for_approval": "[Ẁàìţìñğ ƒøŕ àþþŕøṽàĺ ··················]",
       "waiting_for_approval_address": "[Ẁàìţìñğ ƒøŕ ţĥè àþþŕøṽàĺ àððŕèşş… ìţ àþþèàŕş ĥèŕè ẁĥèñ ţĥè àğèñţ ñèẋţ çøññèçţş — şèñð àñý çĥàţ ɱèşşàğè ţø şţàŕţ àùţĥøŕìžàţìøñ ñøẁ. ·······································]"
     },
+    "discoverPage": {
+      "subtitle": "[Ɓŕøẁşè àñð ìñşţàĺĺ àğèñţìç àþþş ················]",
+      "title": "[Ðìşçøṽèŕ ············]"
+    },
     "mcpManagement": {
       "assessment": {
         "col_assessment": "[Àşşèşşɱèñţ ···············]",
@@ -12822,8 +12821,9 @@
   "nav": {
     "agent_capabilities": "[Àğèñţ Çàþàƀìĺìţìèş ················]",
     "artifacts": "[Àŕţìƒàçţş ··············]",
-    "explore": "[Èẋþĺøŕè ···········]",
+    "discover": "[Ðìşçøṽèŕ ············]",
     "knowledge": "[Ķñøẁĺèðğè ··············]",
+    "library": "[Ĺìƀŕàŕý ···········]",
     "notifications": "[Ñøţìƒìçàţìøñş ············]",
     "schedule": "[Şçĥèðùĺè ············]",
     "sessions": "[Şèşşìøñş ············]",
