+++ website/src/i18n/locales/en.json
@@ -13,8 +13,6 @@
     "discord_community": "Discord community",
     "dismiss": "Dismiss",
     "dsk": "DSK",
-    "explore": "Explore",
-    "explore_apps": "Explore Apps",
     "k_search_for_anything": "⌘K — Search for anything…",
     "kiro_credit_usage_checking": "Kiro credit usage — checking…",
     "kiro_credit_usage_checking_2": "Kiro credit usage — checking",
@@ -5230,12 +5228,11 @@
       "loading_app": "Loading app…"
     },
     "appsPage": {
-      "add_an_app_source_gear_icon_above_or_install_fro": "Add an app source (gear icon above) or install from a local path.",
+      "add_an_app_source_gear_icon_above_or_install_fro": "Add an app source (Sources, top right) or install from a local path.",
       "agent_one": "{{count}} agent",
       "agent_other": "{{count}} agents",
       "app_one": "{{count}} app",
       "app_other": "{{count}} apps",
-      "apps": "Apps",
       "available": "available",
       "cancel": "Cancel",
       "category": "Category",
@@ -5243,7 +5240,6 @@
       "cron_job_one": "{{count}} cron job",
       "cron_job_other": "{{count}} cron jobs",
       "dependencies": "Dependencies:",
-      "discover_install_and_manage_agentic_apps": "Discover, install, and manage agentic apps",
       "dismiss_message": "Dismiss message",
       "keep_app_data": "Keep app data",
       "keep_this_dependency": "Keep this dependency",
@@ -5267,7 +5263,7 @@
       "update_one": "{{count}} update",
       "update_other": "{{count}} updates",
       "v": "v",
-      "installed_app_find_in_library_and_enable": "Installed {{name}}. Turn it on below to enable it.",
+      "installed_app_find_in_library_and_enable": "Installed {{name}}. Enable it in Library.",
       "synced_from_its_source_directory": "Synced {{name}} from its source directory.",
       "updated_from_the_registry": "Updated {{name}} from the registry."
     },
