+++ website/src/i18n/locales/ja.json
@@ -34,8 +34,6 @@
     "dismiss": "閉じる",
     "dsk": "DSK",
     "expand_sidebar": "サイドバーを展開",
-    "explore": "探索",
-    "explore_apps": "アプリを探索",
     "gateway_connected": "ゲートウェイに接続済み",
     "gateway_offline": "ゲートウェイがオフライン",
     "gateway_offline_reconnecting": "ゲートウェイがオフライン — 再接続中",
@@ -5175,7 +5173,6 @@
         },
         "pagesProvider": {
           "developer": "デベロッパー",
-          "explore": "探索",
           "hooks": "フック",
           "kirocrew_agents": "{{productName}} エージェント",
           "logs": "ログ",
@@ -7316,8 +7313,9 @@
   "nav": {
     "agent_capabilities": "エージェント機能",
     "artifacts": "アーティファクト",
-    "explore": "探索",
+    "discover": "発見",
     "knowledge": "ナレッジ",
+    "library": "ライブラリ",
     "notifications": "通知",
     "schedule": "スケジュール",
     "sessions": "セッション",
@@ -7534,11 +7532,10 @@
     },
     "appsPage": {
       "action_failed": "{{name}} の{{action}}に失敗しました",
-      "add_an_app_source_gear_icon_above_or_install_fro": "アプリソースを追加（上の歯車アイコン）またはローカルパスからインストールします。",
+      "add_an_app_source_gear_icon_above_or_install_fro": "アプリソースを追加（右上の「ソース」）するか、ローカルパスからインストールしてください。",
       "agent_other": "{{count}} エージェント",
       "all_apps": "すべてのアプリ",
       "app_other": "{{count}} アプリ",
-      "apps": "アプリ",
       "available": "利用可能",
       "built_in_kirocrew": "ビルトイン · kirocrew",
       "cancel": "キャンセル",
@@ -7546,16 +7543,14 @@
       "confirm_uninstall": "アンインストールを確認",
       "cron_job_other": "{{count}} cron ジョブ",
       "dependencies": "依存関係：",
-      "discover": "発見",
-      "discover_install_and_manage_agentic_apps": "エージェンティックアプリの発見、インストール、管理",
       "dismiss_message": "メッセージを非表示",
       "failed_to_enable": "{{name}} の有効化に失敗しました",
       "failed_to_load_apps": "アプリの読み込みに失敗",
       "failed_to_load_registry": "レジストリの読み込みに失敗",
       "failed_to_uninstall": "{{name}} のアンインストールに失敗しました",
       "failed_to_update": "更新に失敗しました：{{names}}",
-      "find_apps_in_the_discover_tab_or_install_from_a": "[Discover] タブでアプリを見つけるか、上のソース歯車からローカルパスを使ってインストールしてください。",
-      "hidden_you_can_re_enable_it_from_the_discover_ta": "無効化しました。[Discover] タブから再度有効化できます。",
+      "find_apps_in_the_discover_tab_or_install_from_a": "アプリは「見つける」ページで探すか、そのページの「ソース」からローカルパスでインストールできます。",
+      "hidden_you_can_re_enable_it_from_the_discover_ta": "無効にしました。「見つける」ページから再有効化できます。",
       "keep_app_data": "アプリデータを保持",
       "keep_dependency": "依存関係 {{name}} を保持",
       "keep_this_dependency": "この依存関係を保持",
@@ -7564,6 +7559,7 @@
       "kirocrew_registry": "{{productName}} レジストリ",
       "library": "ライブラリ",
       "loading_apps": "アプリを読み込み中…",
+      "manage_your_installed_apps": "インストール済みアプリを管理",
       "name": "名前",
       "no_apps_available": "利用可能なアプリがありません",
       "no_apps_installed_yet": "まだアプリがインストールされていません",
@@ -7590,7 +7586,7 @@
       "update_other": "{{count}} 更新",
       "updating_progress": "更新中 {{done}}/{{total}}…",
       "v": "v",
-      "installed_app_find_in_library_and_enable": "{{name}} をインストールしました。下でオンにして有効化してください。",
+      "installed_app_find_in_library_and_enable": "{{name}} をインストールしました。ライブラリで有効にしてください。",
       "synced_from_its_source_directory": "{{name}} をソースディレクトリから同期しました。",
       "updated_app_other": "{{count}} アプリを更新しました。",
       "updated_from_the_registry": "{{name}} をレジストリから更新しました。"
@@ -9226,6 +9222,10 @@
           "label": "テレメトリー"
         }
       }
+    },
+    "discoverPage": {
+      "subtitle": "エージェンティックアプリの閲覧とインストール",
+      "title": "発見"
     },
     "embedSettingsPage": {
       "back": "戻る",
