+++ website/src/test/AppsPageW3Coverage.test.tsx
@@ -99,7 +99,9 @@ vi.mock('../components/SimpleSelect', () => ({
   ),
 }))
 
-import AppsPage, { pickFeatured } from '../pages/AppsPage'
+import DiscoverPage from '../pages/apps/DiscoverPage'
+import LibraryPage from '../pages/apps/LibraryPage'
+import { pickFeatured } from '../pages/apps/useAppsData'
 import type { RegistryApp } from '../components/appstore/types'
 
 /** Detail-route probe: reports the app name and the `autoAction` router state. */
@@ -111,13 +113,16 @@ function DetailProbe() {
   )
 }
 
-function renderPage() {
+/** Mounts the split pages under their real routes. Library tests mount at
+ *  '/apps/library' directly — the in-page Library tab is gone (PR1 split). */
+function renderPage(initialRoute = '/apps') {
   const qc = new QueryClient({ defaultOptions: { queries: { retry: false, gcTime: 0 } } })
   return render(
     <QueryClientProvider client={qc}>
-      <MemoryRouter initialEntries={['/apps']}>
+      <MemoryRouter initialEntries={[initialRoute]}>
         <Routes>
-          <Route path="/apps" element={<AppsPage />} />
+          <Route path="/apps" element={<DiscoverPage />} />
+          <Route path="/apps/library" element={<LibraryPage />} />
           <Route path="/apps/detail/:name" element={<DetailProbe />} />
           <Route path="/apps/:name" element={<DetailProbe />} />
         </Routes>
@@ -210,7 +215,8 @@ async function browseRow(name: string) {
   return rows[rows.length - 1]
 }
 
-const goLibrary = () => fireEvent.click(screen.getByText('Library'))
+/** Mount the Library route directly — the split removed the in-page tab. */
+const renderLibrary = () => renderPage('/apps/library')
 
 describe('AppsPage — Discover navigation', () => {
   it('Cmd-clicks a row into a new tab instead of routing in place', async () => {
@@ -328,7 +334,7 @@ describe('AppsPage — query failures and empty states', () => {
     listRegistry.mockResolvedValue({ apps: [] })
     renderPage()
     expect(await screen.findByText('No apps available')).toBeInTheDocument()
-    expect(screen.getByText('Add an app source (gear icon above) or install from a local path.')).toBeInTheDocument()
+    expect(screen.getByText('Add an app source (Sources, top right) or install from a local path.')).toBeInTheDocument()
   })
 
   it('shows the no-match state when the search excludes every row', async () => {
@@ -342,28 +348,22 @@ describe('AppsPage — query failures and empty states', () => {
   it('shows the empty Library state when no app is installed', async () => {
     listApps.mockResolvedValue([])
     listRegistry.mockResolvedValue({ apps: [] })
-    renderPage()
-    await screen.findByText('No apps available')
-    goLibrary()
+    renderLibrary()
     expect(await screen.findByText('No apps installed yet')).toBeInTheDocument()
   })
 
   it('shows the no-match Library state when the search excludes every installed app', async () => {
-    renderPage()
-    await catalogReady()
-    goLibrary()
+    renderLibrary()
     await screen.findByText('Slack inbox manager.')
-    fireEvent.change(screen.getByLabelText('Search apps'), { target: { value: 'nothing-matches-this' } })
+    fireEvent.change(screen.getByLabelText('Search library'), { target: { value: 'nothing-matches-this' } })
     expect(await screen.findByText('No matching apps')).toBeInTheDocument()
     expect(screen.getByText('Try a different search term')).toBeInTheDocument()
   })
 
   it('matches an installed app by its manifest tags', async () => {
-    renderPage()
-    await catalogReady()
-    goLibrary()
+    renderLibrary()
     await screen.findByText('Slack inbox manager.')
-    fireEvent.change(screen.getByLabelText('Search apps'), { target: { value: 'inbox' } })
+    fireEvent.change(screen.getByLabelText('Search library'), { target: { value: 'inbox' } })
     // The tag match keeps the row; nothing falls through to the empty state.
     await waitFor(() => expect(screen.queryByText('No matching apps')).toBeNull())
     expect(screen.getByText('Slack inbox manager.')).toBeInTheDocument()
@@ -372,9 +372,7 @@ describe('AppsPage — query failures and empty states', () => {
 
 describe('AppsPage — Library actions', () => {
   it('opens the app at its AppHost route from the card, and the detail page from its name', async () => {
-    renderPage()
-    await catalogReady()
-    goLibrary()
+    renderLibrary()
     // Open must resolve through the same appNavTarget derivation the sidebar and
     // command palette use: a third-party (non-builtin) app is AppHost-routed at
     // /apps/<name>, NOT its raw manifest page route (which only a native builtin
@@ -385,9 +383,7 @@ describe('AppsPage — Library actions', () => {
   })
 
   it('routes to the detail page when the card name is clicked', async () => {
-    renderPage()
-    await catalogReady()
-    goLibrary()
+    renderLibrary()
     fireEvent.click(await screen.findByRole('button', { name: 'Secretary' }))
     const probe = await screen.findByTestId('detail-route')
     expect(probe).toHaveAttribute('data-path', '/apps/detail/secretary')
@@ -396,39 +392,31 @@ describe('AppsPage — Library actions', () => {
 
   it('toasts after disabling a builtin and clears the toast on dismiss', async () => {
     listApps.mockResolvedValue([{ ...BUILTIN_OFF, enabled: true }])
-    renderPage()
-    await catalogReady()
-    goLibrary()
+    renderLibrary()
     fireEvent.click(await screen.findByRole('button', { name: 'Disable' }))
     await waitFor(() => expect(disableApp).toHaveBeenCalledWith('pets'))
-    expect(await screen.findByText('Disabled. You can re-enable it from the Discover tab.')).toBeInTheDocument()
+    expect(await screen.findByText('Disabled. You can re-enable it from the Discover page.')).toBeInTheDocument()
     fireEvent.click(screen.getByRole('button', { name: 'Dismiss message' }))
     await waitFor(() =>
-      expect(screen.queryByText('Disabled. You can re-enable it from the Discover tab.')).toBeNull())
+      expect(screen.queryByText('Disabled. You can re-enable it from the Discover page.')).toBeNull())
   })
 
   it('reports a failed disable with the action-failed message', async () => {
     disableApp.mockRejectedValue({})
-    renderPage()
-    await catalogReady()
-    goLibrary()
+    renderLibrary()
     fireEvent.click(await screen.findByRole('button', { name: 'Disable' }))
     expect(await screen.findByText('Failed to disable secretary')).toBeInTheDocument()
   })
 
   it('names the apps that failed during Update All', async () => {
     updateApp.mockRejectedValue(new Error('clone failed'))
-    renderPage()
-    await catalogReady()
-    goLibrary()
+    renderLibrary()
     fireEvent.click(await screen.findByRole('button', { name: 'Update All' }))
     expect(await screen.findByText('Failed to update: secretary')).toBeInTheDocument()
   })
 
   it('reports success after Update All and shows no error', async () => {
-    renderPage()
-    await catalogReady()
-    goLibrary()
+    renderLibrary()
     fireEvent.click(await screen.findByRole('button', { name: 'Update All' }))
     expect(await screen.findByText('Updated 1 app.')).toBeInTheDocument()
     expect(screen.queryByText(/Failed to update/)).toBeNull()
@@ -437,9 +425,7 @@ describe('AppsPage — Library actions', () => {
 
 describe('AppsPage — uninstall dialog', () => {
   const openDialog = async () => {
-    renderPage()
-    await catalogReady()
-    goLibrary()
+    renderLibrary()
     fireEvent.click(await screen.findByRole('button', { name: 'Uninstall' }))
     return screen.findByRole('dialog', { name: 'Confirm uninstall' })
   }
@@ -619,9 +605,7 @@ describe('AppsPage — editorial layer wiring', () => {
 
 describe('AppsPage — Library enable and update', () => {
   it('Update on a Library card routes to the streaming detail page', async () => {
-    renderPage()
-    await catalogReady()
-    goLibrary()
+    renderLibrary()
     fireEvent.click(await screen.findByRole('button', { name: 'Update' }))
     const probe = await screen.findByTestId('detail-route')
     expect(probe).toHaveAttribute('data-path', '/apps/detail/secretary')
@@ -639,9 +623,7 @@ describe('AppsPage — Library enable and update', () => {
       source: '/home/u/apps/orchestrator-switch', origin: 'local',
     }])
     listRegistry.mockResolvedValue({ apps: [] })
-    renderPage()
-    await screen.findByText('No apps available')
-    goLibrary()
+    renderLibrary()
     fireEvent.click(await screen.findByRole('button', { name: 'Sync' }))
     await waitFor(() => expect(updateApp).toHaveBeenCalledWith('orchestrator-switch'))
     expect(screen.queryByTestId('detail-route')).toBeNull()
@@ -657,9 +639,7 @@ describe('AppsPage — Library enable and update', () => {
       source: '/home/u/apps/orchestrator-switch', origin: 'local',
     }])
     listRegistry.mockResolvedValue({ apps: [] })
-    renderPage()
-    await screen.findByText('No apps available')
-    goLibrary()
+    renderLibrary()
     fireEvent.click(await screen.findByRole('button', { name: 'Sync' }))
     expect(await screen.findByText(/Synced Orchestrator Switch from its source directory/)).toBeInTheDocument()
   })
@@ -671,9 +651,7 @@ describe('AppsPage — Library enable and update', () => {
     }])
     listRegistry.mockResolvedValue({ apps: [] })
     updateApp.mockRejectedValue(new Error('source path no longer exists'))
-    renderPage()
-    await screen.findByText('No apps available')
-    goLibrary()
+    renderLibrary()
     fireEvent.click(await screen.findByRole('button', { name: 'Sync' }))
     expect(await screen.findByText('source path no longer exists')).toBeInTheDocument()
   })
@@ -681,9 +659,7 @@ describe('AppsPage — Library enable and update', () => {
   it('still routes a registry-sourced app at the registry when it has no update', async () => {
     listApps.mockResolvedValue([{ ...SECRETARY, source: 'registry:secretary' }])
     listRegistry.mockResolvedValue({ apps: [] })
-    renderPage()
-    await screen.findByText('No apps available')
-    goLibrary()
+    renderLibrary()
     fireEvent.click(await screen.findByRole('button', { name: 'Sync' }))
     const probe = await screen.findByTestId('detail-route')
     expect(probe).toHaveAttribute('data-auto', 'update')
@@ -692,9 +668,7 @@ describe('AppsPage — Library enable and update', () => {
 
   it('enables a switched-off installed app from the Library', async () => {
     listApps.mockResolvedValue([{ ...SECRETARY, enabled: false }])
-    renderPage()
-    await catalogReady()
-    goLibrary()
+    renderLibrary()
     fireEvent.click(await screen.findByRole('button', { name: 'Enable' }))
     await waitFor(() => expect(enableApp).toHaveBeenCalledWith('secretary'))
   })
@@ -710,9 +684,7 @@ describe('AppsPage — Library enable and update', () => {
     }])
     listRegistry.mockResolvedValue({ apps: [] })
     enableApp.mockRejectedValue({ code: 'app_execution_denied' })
-    renderPage()
-    await screen.findByText('No apps available')
-    goLibrary()
+    renderLibrary()
     fireEvent.click(await screen.findByRole('button', { name: 'Enable' }))
     expect(await screen.findByText('Trust Local App to run its own code?')).toBeInTheDocument()
     expect(screen.queryByText('https://github.com/z/secretary')).toBeNull()
@@ -722,19 +694,15 @@ describe('AppsPage — Library enable and update', () => {
 describe('AppsPage — toast expiry', () => {
   it('the disable toast clears itself after four seconds', async () => {
     listApps.mockResolvedValue([{ ...BUILTIN_OFF, enabled: true }])
-    renderPage()
-    await catalogReady()
-    goLibrary()
+    renderLibrary()
     fireEvent.click(await screen.findByRole('button', { name: 'Disable' }))
-    await screen.findByText('Disabled. You can re-enable it from the Discover tab.')
+    await screen.findByText('Disabled. You can re-enable it from the Discover page.')
     await act(async () => { vi.advanceTimersByTime(4000) })
-    expect(screen.queryByText('Disabled. You can re-enable it from the Discover tab.')).toBeNull()
+    expect(screen.queryByText('Disabled. You can re-enable it from the Discover page.')).toBeNull()
   })
 
   it('the Update All toast clears itself after four seconds', async () => {
-    renderPage()
-    await catalogReady()
-    goLibrary()
+    renderLibrary()
     fireEvent.click(await screen.findByRole('button', { name: 'Update All' }))
     await screen.findByText('Updated 1 app.')
     await act(async () => { vi.advanceTimersByTime(4000) })
